+++ src/server/routers/lambda/__tests__/aiModel.test.ts
@@ -4,7 +4,7 @@ import { AiModelModel } from '@/database/models/aiModel';
 import { UserModel } from '@/database/models/user';
 import { AiInfraRepos } from '@/database/repositories/aiInfra';
 
-import { aiModelRouter } from './aiModel';
+import { aiModelRouter } from '../aiModel';
 
 vi.mock('@/database/models/aiModel');
 vi.mock('@/database/models/user');
