+++ src/server/routers/lambda/__tests__/aiProvider.test.ts
@@ -6,7 +6,7 @@ import { getServerGlobalConfig } from '@/server/globalConfig';
 import { KeyVaultsGateKeeper } from '@/server/modules/KeyVaultsEncrypt';
 import { AiProviderDetailItem, AiProviderRuntimeState } from '@/types/aiProvider';
 
-import { aiProviderRouter } from './aiProvider';
+import { aiProviderRouter } from '../aiProvider';
 
 vi.mock('@/server/globalConfig');
 vi.mock('@/server/modules/KeyVaultsEncrypt');
