+++ src/server/routers/lambda/__tests__/generation.test.ts
@@ -6,7 +6,7 @@ import { GenerationModel } from '@/database/models/generation';
 import { FileService } from '@/server/services/file';
 import { AsyncTaskStatus } from '@/types/asyncTask';
 
-import { generationRouter } from './generation';
+import { generationRouter } from '../generation';
 
 vi.mock('@/database/models/asyncTask');
 vi.mock('@/database/models/generation');
